+++ src/main.rs
@@ -11,8 +11,7 @@ mod gpx_writer;
 mod ipc_handler;
 mod map_data;
 mod map_data_cache;
-mod osm_data_reader;
-mod osm_json_parser;
+mod osm_data;
 mod result_writer;
 mod router;
 mod router_runner;
