+++ src/router_runner.rs
@@ -6,12 +6,12 @@ use clap::Parser;
 use serde::{Deserialize, Serialize};
 use tracing::{info, trace};
 
+use crate::osm_data::DataSource;
 use crate::{
     debug::writer::DebugWriter,
     ipc_handler::{IpcHandler, IpcHandlerError, ResponseMessage, RouteMessage, RouterResult},
     map_data::graph::MapDataGraph,
     map_data_cache::{MapDataCache, MapDataCacheError},
-    osm_data_reader::DataSource,
     result_writer::{DataDestination, ResultWriter, ResultWriterError},
     router::{
         generator::{Generator, RouteWithStats},
